+++ Editor/Artifice_ArtificeListView/Artifice_VisualElement_ListView.cs
@@ -22,7 +22,6 @@ namespace ArtificeToolkit.Editor
             
             // Create property's GUI with ArtificeDrawer
             var propertyField = ArtificeDrawer.CreatePropertyGUI(property, shouldForceArtifice);
-            propertyField = ArtificeDrawer.CreateCustomAttributesGUI(property, propertyField, ChildrenInjectedCustomAttributes);
             propertyField.AddToClassList("property-field");
 
             return propertyField;