+++ Runtime/CustomAttributes/MinValueAttribute.cs
@@ -1,7 +1,7 @@
 namespace ArtificeToolkit.Attributes
 {
     /// <summary> Bounds the property's minimum value to a given value </summary>
-    public class MinValueAttribute : CustomAttribute
+    public class MinValueAttribute : ValidatorAttribute
     {
         public float Value;